requirements.txt
# ========================
# Core Django & Extensions
# ========================
Django==4.2.14
djangorestframework==3.14.0
django-extensions==3.2.3
django-environ==0.11.2
django-widget-tweaks==1.5.0
django-crontab==0.7.1

# ========================
# WSGI / ASGI Servers
# ========================
gunicorn==21.2.0
uvicorn[standard]==0.30.0

# ========================
# Security & TLS
# ========================
cryptography==41.0.5
pyOpenSSL==24.1.0
django-cors-headers==4.4.0
whitenoise==6.7.0

# ========================
# Network & Reputation
# ========================
ipwhois==1.2.0
requests==2.32.3
httpx==0.27.0

# ========================
# Validation
# ========================
jsonschema==4.23.0

# ========================
# AI & Data
# ========================
pandas==2.2.2
scikit-learn==1.5.0

# ========================
# Database
# ========================
psycopg2-binary==2.9.9

# ========================
# Queue / Redis (Optional)
# ========================
redis==5.0.8
celery==5.4.0

# ========================
# Excel Handling
# ========================
openpyxl==3.1.5

# ========================
# Env Variables
# ========================
python-dotenv==1.0.1

# ========================
# Logging
# ========================
prometheus_client==0.19.0

geoip2>=4.8.0

